STM: new API async_query
When async_proofs_always_delegate is on, park proof workers and
dispatch to them queries.  This flips the current way of printing
goals in PIDE base UIs: it is not the worker that prints all goals
while coomputing the states (and the dies), it is the UI that asks
for them when they are under the eyes of the user.