proof_global: make it possible to call close_proof in a worker
Given that the proof state contains a callback (a terminator)
that is not sent (dropped by the ephemeron mechanism at marshall
time) de-referencing the ephemeron during this function makes it
impossible to call it in a worker.  Now the worker can call the
function and replace the terminator with a good one.