Add extraction to JSON.
This patch allows Coq terms to be extracted into the widely used JSON
format.  This is useful in at least two cases:

  - One might want to manipulate Coq values outside of Coq, but without
    being forced to use one of the three existing extraction languages
    (OCaml, Haskell, or Scheme), and without having to compile Coq's
    extracted result.  This is especially useful when a Coq evaluation
    produces some data structure that needs to be moved out of Coq.
    Having to invoke an OCaml/Haskell/Scheme compiler just to get a
    data structure out of Coq is somewhat awkward.

  - One might want to experiment with extracting Coq code into other
    languages (Go, Javascript, etc), without having to write the whole
    extraction logic in OCaml and recompile Coq's extraction plugin
    each time.  This makes it easy to quickly prototype extraction
    in any language, without having to build Coq from source.

Extraction to JSON is implemented by adding the JSON "pseudo-language"
to the extraction facility.  Thus, one can extract the JSON encoding
of a single term using:

  Extraction Language JSON.
  Extraction qualid.

and extract an entire Coq library "ident" into "ident.json" using:

  Extraction Language JSON.
  Extraction Library ident.

Nota (Pierre Letouzey) : this is an updated version of the original
PullRequest, updated to match recent changes in trunk