When looking for restrictions in ?n=?p problems, keep the type of let-bindings.
Bindings of the form [let x:T := M] are unfolded into [(M:T)], so that
occur-check is done in [T] as well as in [M] (except when [M] is a
variable, where it is hopefully not necessary).

This is a way to fix #4062 (evars with type depending on themselves).

The fix modifies the alias map (make_alias_map) but it should behave
the same at all places using alias maps other than
has_constrainable_free_vars.