More resilient implementation of Print Assumptions.
Instead of registering all the transitive dependencies of a term in one go,
we rather recursively build the graph of direct dependencies of this term.
This is finer-grained and offers a better API.

The traversal now uses the standard term fold operation, and also registers
inductives and constructors encountered in the traversal.