Fixed #4138. In emacs mode Set/Unset does not print the goal anymore.
In PG there are shortcuts that perform on the fly "Set Printing
All"/"Unset Printing All" in pg. For example if you type C-u before
some shortcut for print (check/About/Show), it performs:

Set Printing All.
Print foo.
Unset Printing All.

But if the "Unset" prints a goal, then pg interprets the output of
Print as a command applied on a previous line, and thus hides it.

So for emacs mode I would prefer no goal printing when options are
set/unset. In the situation where this should be done (when setting
durably the option probably), I'd rather program a "Show" explicitely.