STM: change how proof mode switching commands are handled (decl_mode)
This is likely to make nested proofs containing proof modes switch
broken, but fixes the problems Arnaud has in his branch.

It is possible that the classification function we have today
is not fine grained enough.
If a command that changes the proof mode has as the only global
effect changing the proof mode, then the current code is fine.
If it has a more global effect that persists after the proof is over
but has no impact on the proof itself, then the old code is fine.

As far as I can get, the proof mode switching commands have
a global effect (changing parser) but also a proof effect
(un/focusing).  We don't have a classification for these.

Today a command having a global effect is played twice:
on the proof branch an on master.  Of course if such command
cannot work on master (where there is no finished proof for
example) we need a special treatment for it.