A more reasonable implementation of Loadpath.
The new behaviour is simple: either a path is in the loadpaths or it is not.
No more wild expansions of paths!

This should not affect -R and -Q, but it does change the semantics of -I -as.
Still, there are no more users of it and it only does so in a subtle way.