CoqIDE: simpler way of reopening/reclosing a proof (Close: 4168)
No "read-only" terminator.  If no terminator is present the UI
complains.  If the terminator is different, STM warns but
continues.  The STM warns that the "check the document" button
will not honor the terminator change, and what to do to avoid
that.

Technically, one cannot turn (a posteriori) an axiom into a theorem
and vice versa.  Could be done, but not with a small patch.