Move ad-hoc declaration in constrextern to expected form in coqlib
This works, but will have to become a plugin, to avoid making the standard
library depend on the existence of a constant for presentation.  However,
my fear that the approach would have bad computational behaviour is
irrelevant: augmented terms are seen only by the user, after
externalization; terms in defined constants and goals are not augmented.