Implement [Set Printing Coercions Mode "Cast"], but is a failure
As the generated presentation terms contain holes (_), not genuine types,
the effective target type of the applied coercion is not included in the
generated terms and thus cannot be displayed by the dedicated notation.
Instead, a hole is displayed, leading to “a ↑ _”.  The test file breaks, as
it contains the wanted output, not the actual one.