Fixing bug #3736 (anomaly instead of error/warning/silence on decidability scheme).
Not clear to me why it is not a warning (in verbose mode) rather than
silence when a scheme supposed to be built automatically cannot be
built, as in:

Set Decidable Equality Schemes.
Inductive a := A with b := B.

which could explain why a_beq and a_eq_dec as well as b_beq and
b_eq_dec are not built.