Make coq-tex aware of lines ending with "}", so as to fix the FAQ.
This is only a heuristic and it might cause the tool to become awfully
confused if a line ends with "}" yet this is not the end of a tactic
block. Fixing it would require a full-blown Coq parser inside coq-tex.

Example of crazy output:

Coq < Goal { True }
Coq < 1 subgoal

  ============================
   {True} + {False}

Coq <    + { False }.