STM: save a full state for queries.
In PIDE based UIs queries can be delegated too, hence to speed up
things I was saving a shallow state.  Unfortunately a shallow state
breaks section/modules commands, and a query can be the last entry
of a section/module.  (A shallow state does essentially drop the libstack).

The easy solution is to save a complete state.
A better one would be to refine the static analysis of the document and
decide which kind of saved state one needs.