Fixed (and changed) infoH.
The detection of new hypothesis was bugged.

Now infoH behaves like "Show Intros": it performs tac, grab
information on hypothesis names but let the state unchanged.

FTR: infoH is fundamentally unable to be correct in presence of
tactics that delete hypothesis and reuse there names. Like destruct or
induction. Fortunately destruct and induction now come with a variant
asking that the hypothesis is not deleted. To guess for the right
as-close for [induction H], do [infoH induction !H]. This will not
create the same names as induction would have by itself but at least
there will be the right number of hypothesis.