CLEANUP: removing an unnecessary field of a variant
According to 'parsing/g_vernac.ml4' file, the first field of the

  Vernacexpr.VernacDeclareTacticDefinition

variant is always set to 'true'.
The user has no way to set that field to 'false'.
That value is then just copied around until it is discarded.

Since it currently does not convey any meaningful information,
I've decided to delete it.