CLEANUP: Vernacexpr.vernac_expr
Originally, "VernacTime" and "VernacRedirect" were defined like this:

      type vernac_expr =
      ...
      | VernacTime of vernac_list
      | VernacRedirect of string * vernac_list
      ...

where

      type vernac_list = located_vernac_expr list

Currently, that list always contained one and only one element.
So I propose changing the definition of these two variants in the following way:

      | VernacTime of located_vernac_expr
      | VernacRedirect of string * located_vernac_expr

which covers all our current needs and enforces the invariant
related to the number of commands that are part of the
"VernacTime" and "VernacRedirect" variants.